refactor: Reduce code duplication in WACI Issuance tests
Refactored the WACI Issuance tests using V1 and V2 so that there's less code duplication.
